backend/pyproject.toml
[project]
name = "marble-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "modal>=0.73",
    "torch>=2.3.1",
    "torchvision>=0.18.1",
    "numpy>=1.26",
    "Pillow>=10.3",
    "huggingface_hub",
    "safetensors",
    "roma",
    "einops",
    "opencv-python-headless",
    "scipy",
    "trimesh",
    "fastapi",
    "uvicorn",
    "websockets",
    "transformers>=4.46",
    "open3d",
    "shapely",
]

[dependency-groups]
dev = ["ruff>=0.11", "ty>=0.0.18"]

[tool.ruff]
target-version = "py311"
line-length = 99

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
    "TCH",   # type-checking imports
    "RUF",   # ruff-specific
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["spatiallm", "streamvggt"]

[tool.ty.environment]
python-version = "3.11"
